fix check user balances
